主体基本完成，尚未测试

+++ Client/mSystem.java
@@ -22,7 +22,7 @@ import java.util.LinkedList;
  * @Date 2019/1/19 17:17
  **/
 public class mSystem extends JFrame{
-    public JPanel SystemPanel;
+    private JPanel SystemPanel;
     private JButton exit;
     private JButton edit;
     private JButton delete;
@@ -31,60 +31,67 @@ public class mSystem extends JFrame{
     private JTree information;
     private JButton refresh;
     private JTextField searchField;
+    private JFrame frame;
+
+    DefaultMutableTreeNode rootNode = null;
+    DefaultMutableTreeNode departmentNode0 = null;
+    DefaultMutableTreeNode departmentNode1 = null;
+    DefaultMutableTreeNode departmentNode2 = null;
+    DefaultMutableTreeNode departmentNode3 = null;
+    DefaultMutableTreeNode departmentNode4 = null;
 
     private dbOperate db = null;
-    private dbAuth user = null;
+    private dbAuth admin = null;
 
     public mSystem(dbOperate db) {
         this.db = db;
+        admin = db.getAdmin();
 
         //退出按钮
         exit.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
-                dispose();
+                frame.dispose();
             }
         });
         //添加按钮
         add.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
-                Add add = new Add(new dbUser());
-                add.setDb(db);
-                JFrame frame = new JFrame("添加");
-                frame.setContentPane(add.AddPanel);
-                frame.setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
-                frame.setLocation(500,500);
-                frame.setSize(500,400);
-                frame.pack();
-                frame.setVisible(true);
+                if (admin.getLevel()>=1) {
+                    Add add = new Add(new dbUser());
+                    add.setDb(db);
+                    add.run("添加");
+//                    init(1);
+                    frame.dispose();
+                } else {
+                    JOptionPane.showMessageDialog(null, "当前登录的账户没有权限进行添加操作", "警告", JOptionPane.PLAIN_MESSAGE);
+                }
             }
         });
         //刷新按钮
         refresh.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
-//                init();
-                information.updateUI();
+                init(1);
             }
         });
         //修改
         edit.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
-                DefaultMutableTreeNode node = (DefaultMutableTreeNode) information.getLastSelectedPathComponent();
-                System.out.println("修改的节点：" + information.getSelectionPath());
-                if (node.getUserObject() instanceof dbUser) {
-                    dbUser obj = (dbUser) node.getUserObject();
-                    Add add = new Add(obj);
-                    add.setDb(db);
-                    JFrame frame = new JFrame("修改");
-                    frame.setContentPane(add.AddPanel);
-                    frame.setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
-                    frame.setLocation(500,500);
-                    frame.setSize(500,400);
-                    frame.pack();
-                    frame.setVisible(true);
+                if (admin.getLevel()>=1) {
+                    DefaultMutableTreeNode node = (DefaultMutableTreeNode) information.getLastSelectedPathComponent();
+                    System.out.println("修改的节点：" + information.getSelectionPath());
+                    if (node.getUserObject() instanceof dbUser) {
+                        dbUser obj = (dbUser) node.getUserObject();
+                        Add add = new Add(obj);
+                        add.setDb(db);
+                        add.run("修改");
+                    }
+                    init(1);
+                } else {
+                    JOptionPane.showMessageDialog(null, "当前登录的账户没有权限进行修改操作", "警告", JOptionPane.PLAIN_MESSAGE);
                 }
             }
         });
@@ -92,16 +99,20 @@ public class mSystem extends JFrame{
         delete.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
-                DefaultMutableTreeNode node = (DefaultMutableTreeNode) information.getLastSelectedPathComponent();
-                System.out.println("删除的节点：" + information.getSelectionPath());
-                if (node.getUserObject() instanceof dbUser) {
-                    dbUser obj = (dbUser) node.getUserObject();
-                    if (db.delete(obj)) {
-                        JOptionPane.showMessageDialog(null, "删除成功", "成功", JOptionPane.PLAIN_MESSAGE);
-                    } else {
-                        JOptionPane.showMessageDialog(null, "删除失败", "失败", JOptionPane.PLAIN_MESSAGE);
+                if (admin.getLevel()>=2) {
+                    DefaultMutableTreeNode node = (DefaultMutableTreeNode) information.getLastSelectedPathComponent();
+                    System.out.println("删除的节点：" + information.getSelectionPath());
+                    if (node.getUserObject() instanceof dbUser) {
+                        dbUser obj = (dbUser) node.getUserObject();
+                        if (db.delete(obj)) {
+                            JOptionPane.showMessageDialog(null, "删除成功", "成功", JOptionPane.PLAIN_MESSAGE);
+                        } else {
+                            JOptionPane.showMessageDialog(null, "删除失败", "失败", JOptionPane.PLAIN_MESSAGE);
+                        }
                     }
-
+                    init(1);
+                } else {
+                    JOptionPane.showMessageDialog(null, "当前登录的账户没有权限进行删除操作", "警告", JOptionPane.PLAIN_MESSAGE);
                 }
             }
         });
@@ -118,36 +129,58 @@ public class mSystem extends JFrame{
         this.db = db;
     }
 
-    public void setUser(dbAuth user) {
-        this.user = user;
+    public void setAdmin(dbAuth admin) {
+        this.admin = admin;
     }
 
     private void createUIComponents() {
         // TODO: place custom component creation code here
         //创建树形结构
-        init();
-    }
+        rootNode = new DefaultMutableTreeNode("员工信息");
+        departmentNode0 = new DefaultMutableTreeNode("部门A");
+        departmentNode1 = new DefaultMutableTreeNode("部门B");
+        departmentNode2 = new DefaultMutableTreeNode("部门C");
+        departmentNode3 = new DefaultMutableTreeNode("部门D");
+        departmentNode4 = new DefaultMutableTreeNode("部门E");
 
-    public void init() {
-//        String[] department = {"部门A","部门B","部门C","部门D","部门E"};
-
-//        information.clearSelection();
-
-        DefaultMutableTreeNode rootNode = new DefaultMutableTreeNode("员工信息");
-        rootNode.removeAllChildren();
-
-        DefaultMutableTreeNode departmentNode0 = new DefaultMutableTreeNode("部门A");
-        DefaultMutableTreeNode departmentNode1 = new DefaultMutableTreeNode("部门B");
-        DefaultMutableTreeNode departmentNode2 = new DefaultMutableTreeNode("部门C");
-        DefaultMutableTreeNode departmentNode3 = new DefaultMutableTreeNode("部门D");
-        DefaultMutableTreeNode departmentNode4 = new DefaultMutableTreeNode("部门E");
-
+        //创建用户-部门树状图
         rootNode.add(departmentNode0);
         rootNode.add(departmentNode1);
         rootNode.add(departmentNode2);
         rootNode.add(departmentNode3);
         rootNode.add(departmentNode4);
 
+        init(0);
+    }
+
+    //初始化树状图，type为0是初次初始化，1为刷新
+    private void init(int type) {
+//        String[] department = {"部门A","部门B","部门C","部门D","部门E"};
+
+//        information.clearSelection();
+
+//        //创建树形结构
+//        DefaultMutableTreeNode rootNode = new DefaultMutableTreeNode("员工信息");
+//        DefaultMutableTreeNode departmentNode0 = new DefaultMutableTreeNode("部门A");
+//        DefaultMutableTreeNode departmentNode1 = new DefaultMutableTreeNode("部门B");
+//        DefaultMutableTreeNode departmentNode2 = new DefaultMutableTreeNode("部门C");
+//        DefaultMutableTreeNode departmentNode3 = new DefaultMutableTreeNode("部门D");
+//        DefaultMutableTreeNode departmentNode4 = new DefaultMutableTreeNode("部门E");
+//
+//        //创建用户-部门树状图
+//        rootNode.add(departmentNode0);
+//        rootNode.add(departmentNode1);
+//        rootNode.add(departmentNode2);
+//        rootNode.add(departmentNode3);
+//        rootNode.add(departmentNode4);
+
+        if (type == 1) {
+            departmentNode0.removeAllChildren();
+            departmentNode1.removeAllChildren();
+            departmentNode2.removeAllChildren();
+            departmentNode3.removeAllChildren();
+            departmentNode4.removeAllChildren();
+        }
 
         LinkedList<Object> result = db.Select("user", "Disable", "0");
 
@@ -178,9 +211,14 @@ public class mSystem extends JFrame{
             result.removeFirst();
         }
 
-        information = new JTree(rootNode);
-        information.setShowsRootHandles(true);
-        information.setEditable(false);
+        if (type == 0) {
+            information = new JTree(rootNode);
+            information.setShowsRootHandles(true);
+            information.setEditable(true);
+        } else {
+            information.updateUI();
+        }
+
     }
 
     private void findInTree(String str) {
@@ -216,5 +254,15 @@ public class mSystem extends JFrame{
             }
             return null;
         }
+    }
+
+    public void run() {
+        frame = new JFrame("信息管理系统");
+        frame.setSize(1000,800);
+        frame.setContentPane(this.SystemPanel);
+        frame.setDefaultCloseOperation(JFrame.DISPOSE_ON_CLOSE);
+        frame.setLocation(400,400);
+        frame.pack();
+        frame.setVisible(true);
     }
 }
